Stop all sounds within a clip - better solution
Instead of maintaining a Vec of active sound instances in each
DisplayObject, this commit reuses the already-existing sound
tracking mechanism of each backend.
This is both much more simple, and doesn't leak the instances list.

TODO: Consider to require all sound instances to be associated with
a MovieClip. Currently some are not associated with any (i.e. have
None in their clip_id). This would allow to convert clip_id to be
non-Option.