core: Fix getBytesLoaded and getBytesTotal returning inaccurate counts
The implementation of these functions that SWF headers have a constant
size of either 20 or 21, when in fact it can vary. This assumption was
was causing at least one preloader to get stuck forever, because the
wrong number of loaded bytes was being returned.
https://homestuck.net/games/alabaster-silent-hive/argh.swf

Instead, if we just keep track of the uncompressed_length from the swf
header and return it from these functions, the byte count is correct.