desktop: Fix sound envelopes to use proper positions
The points in an SWF sound envelope are always in 44.1KHz. The
desktop audio backend would end up using this points in the sound's
sample rate, often causing the envelope to be delayed.

Now the envelope is applied after resampling, and the positions
are scaled based on the output sample rate.