desktop: Revert cpal initialization on a separate thread
It was initialized on a separate thread since 6178dd9ef1292163b1f81b6c6e9b99e7d1923ccb,
as a workaround for Windows.

But now thanks to https://github.com/RustAudio/cpal/pull/597, this
is no longer needed, and cpal can be initialized on the same thread
as before.