core: When advancing in a goto, make sure the tag stream position is correct.
AVM2 requires that we read the tag stream thrice per frame; and we can't update the position until run_frame. If we run a goto before that, we will run the tags of the previous frame. So we instead move the tag stream position to the next frame so that we can correctly execute the rest of the goto.