web: Add `isPlaying` property to rufflePlayer
This is a small PR, which adds a little property called isPlaying that is true if the movie is playing, and is false if the movie is paused. I needed this for a website I am working on, and I thought I'd submit it here just incase it is helpful.

By the way, I am quite new to this whole project/repo, so if I did anything wrong here please let me know. Thanks!