render: Don't remove entry from bitmap_register for offscreen rendering
It turns out that Flash supports 'recursive' BitmapData rendering
(calling BitampData.draw with a source that requires rendering a Bitmap
that shares the target BitmapData).

This is a first step towards supporting this case. We now need
to wrap the `Texture` and `Buffer` fields in an `Arc`, in order
to construct the new `TextureTarget`.