avm2: Initial incomplete implementation of XML
This is a (very incomplete) implementation of the XML and XMLList
classes. Basic parsing works, along with child and attribute lookup.

Namespacse, mutation, and the vast majority of XML/XMLList methods
are not implemented. I doubt that this is enough to get any games
working, but it should be enough to allow devs to start working
on different parts of the XML implementation in parallel (now that
the basic framework is somewhat in place).