tests: Add support for comparing test output against real `flashplayer`
This is only supported on Linux for now. It should be straightforward
to add macOS support, and Windows support should be possible with
a little effort.

We don't store `flashplayer` in the repository:
I've added a script which downloads it from an external site.

There were two tricky issues that I needed to solve:
* Normally, flashplayer won't exit when it reaches the last frame.
  We could monitor the output and kill flashplayer when we see
  the expected number of trace lines, but this would make us unable
  to detect hangs that occur after the last line of output.
  However we can use `fscommand("quite")` from ActionScript to
  make flashplayer exit. This unfortunately requires modifying
  all of our SWFs, but this can be done gradually.
* Flashplayer always writes its trace output to the same file,
  even if there are multiple instances of the player running.
  In order to be able to run our tests in parallel, I used
  Frida (https://frida.re/) to hook into `flashplayer`, and
  redirect calls to libc `open` to a different file.

This is currently only enabled for the `try_catch` test.