avm2: Ensure that we don't skip running queued frame scripts
Previously, running `enter_frame` twice without calling
`run_frame_scripts` would cause us to skip running framescripts
queued by the first `enter_frame` call. We now panic if this occurs.

Adding in this check caught a bug in Loader, which wasn't skipping
the first 'enterFrame' call for the constructed MovieClip
(this gets treated in the same way as constructing a MovieClip
from ActionScript and adding it to the stage).