wgpu: Discard transparent bitmap pixels in mask when cacheAsBitmap=true
When we draw a mask, we increment the stencil buffer for each pixel
we draw. Flash supports using a Bitmap as a mask, and only considers
pixels with alpha!=0 to be part of the mask when cacheAsBitmap=true
is set on *both* the masker and maskee.

This can be implemented by using 'discard' in our 'bitmap.wgsl'
shader for pixels with 'alpha=0', which prevents the stencil
buffer from getting updated.

This fixes using unlockable pants textures in Fancy Pants World 4
Part 3, which previously rendered as a large rectangle over the player.