core: Stub out SystemCapabilities bitmap and improve default values
The SystemCapabilities bitmap contains the flash api system.capabilities
booleans, which are currently not implemented. Previously, the bitmap
was empty, meaning that all booleans were set to false.

Now, the bitmap values have been stubbed out. If clauses setting the
respective booleans have been added. The actual tests aren't
implemented, but the stubs are filling the bitmap with probable entries
(e.g. most systems have audio encoding, therefore the audio encoding
boolean is now set to true).
This allows the capabilities to be more accurate (while not being
implemented yet) and simplifies the actual implementing, since only the
conditions of the stubs need to be replaced.