core: Add Player RuffleType & Implement system.capabilities.playerType
A RuffleType enum has been added. It has the possible values
DesktopPlayer and WebPlayer. The Player now has a variable ruffle_type
whose type is this enum. The PlayerBuilder sets the ruffle_type of a
Player via conditional compilation.
The SystemProperties struct now receives the ruffle_type at its
creation. It uses it to implement system.capabilities.playerType.
The playerType is set to StandAlone if the desktop player is used and to
PlugIn if the web player is used.

Closes #8309.