core: Implement system.capabilities.os
The SystemProperty constructor now recognises the OS and sets the os
variable. It uses the new get_os helper function, which has multiple
versions using conditional compilation.
If a desktop version is used, it gets the OS using the environment
constant. If Windows is used, it determines the Windows version using
the dependency os_info. If that's not possible, the ver CLI command is
used and the result parsed instead. The Windows version then gets
matched to one of the possible os values.
If the web version is used, it gets the user agent using the dependency
web-sys. The user agent is then used to determine the OS and the
variable is set accordingly.

Co-authored-by: bestlinuxgamers <52172848+bestlinuxgamers@users.noreply.github.com>