core: Implement system.capabilities.cpuArchitecture
The SystemProperty constructor now recognises the CPU architecture and
sets the cpu_architecture variable. It uses the new get_cpu_architecture
helper function, which has multiple versions using conditional
compilation.
If a desktop version is used, it gets the CPU architecture using the
environment constant.
If the web version is used, it gets the user agent string using the
dependency web-sys. The user agent is then used to determine the CPU
architecture. That is a rudimentary implementation since most user
agents don't allow to infer the architecture, but it seems that there
is no more reliable way.