core: Log an error when a font is missing a glyph for a character
My hope is that this never happens for real SWFS (unless Ruffle has
a bug). If this does get hit intentionally in real SWFs, then
it's going to be very verbose (logging each time we render), so
we'll need to adjust this to log only once for a given char.