core: Fully implement system.capabilities.language
Previously, system.capabilities.language has always used the locale on
Windows. However, according to the flash specs, the UI language should
be used on Windows if the flash player version is >= 7.
This has now been implemented. If the OS is Windows and the flash player
version >= 7, the UI language is now retrieved via the Windows API
(using the winapi crate) and used instead of the locale.

Additionally, the language string is now lowercased before parsing and
converting it into the language variables. This way, the language
detection still works if an OS returns the language string in uppercase.

Co-authored-by: bestlinuxgamers <52172848+bestlinuxgamers@users.noreply.github.com>