desktop: Add CFBundleDocumentTypes info to Info.plist
The CFBundleDocumentTypes information has been added to the Info.plist
on macOS.
This part of the Info.plist tells the operating system which file
extensions a program can open.
The added information states that Ruffle can open SWF files and should
be the default program to open an SWF file with.

Additionally, this should cause CFBundleTypeIconFiles to be generated
during the build process.
This means that SWF files will have the Ruffle logo on their file icon.