extension: Log used renderer; Warn if not preferred.
Log which renderer-backend was actually used in the
browser console.

Add an option for toggling whether to warn or not
if the actually used renderer is not the same as
the preferred renderer, warning both in the console
and also with a `window.alert()` popup. This option
is meant first and foremost for Ruffle-developers.