tests: Unconditionally save 'actual.txt' with Ruffle output
When a test has large amounts of output, the diff printed by
Ruffle can be difficult to understand. Also, it can be useful
to see the real floating point values for a passing test with
approximate floating point comparison enabled.

I've modified the test runner to save the actual Ruffle trace output
to 'actual.txt' (in the same directory as 'output.txt'). For simplicity,
this is done regardless of whether or not the test passes or fails.

To prevent messy 'git status' output and accidental commits, I've added
all of the test-generated files ('actual.txt', 'different-*.png', and
'actual-*png') to .gitignore